remove useless code of table buttons
